Merged PR 5611: check if enabled before doing api calls
check if enabled before doing api calls

Cron execution to unsubcribe suppressed contacts should check for api enabled before doing api calls.

Currently it checks for the client but the client is always true - so the check fails to stop execution.

Fix should check if api is enabled for each website that is configured to run the sync.

Related work items: #63854