Merged PR 5835: Support the templates for store level
Because the templates have the code unique we can't save the same template for different stores.

On unmapping we can't trace what template was removed and we need to delete from email_templates table.

To support the email templates for each store we have to add the store code to the template name ex.

'New Account (dotmailer)__2" <--- 2 will mean that this template was created and mapped for the store with id = 2

Related work items: #64486