Merged PR 5936: fix emailcapture to make infinite loop calls
Fix emailcapture to make infinite loop calls.

On checkout page the emailcapture is called multiple/infinite times when the password manager is enabled and the user saved the creds to be inserted.

This should fix the multiple calls and checking the previous post do doesn't duplicate the post send if the email has not changed.

Related work items: #65062