Merged PR 6039: Don't unbind when email is empty, invalid or not changed. If we do, the blur
Don't unbind when email is empty, invalid or not changed. If we do, the blur event is not listened to ever again.

This rolls back to the simpler version we had before. It now saves the last received email, if it's not changed since last blur, then return false as it's likely we're in an infinite loop caused by browser auto fills.

Related work items: #65062