Fixes for newsletter subscription options
When 'Show datafields to customer' is set to false, the request does not contain data_fields. This causes an error when a user updates their additional newsletter subscriptions. The change to getContactDataFields checks if it's set, if not returns an empty array. If it has been set, but is an empty array, it returns an empty array too. This saves an additional call to dotmailer.

As per github 502, we weren't correctly updating preferences. Now we iterate over the configured additional address books and maintain those mappings. We don't touch other address book subscriptions.

https://github.com/dotmailer/dotmailer-magento2-extension/issues/501
https://github.com/dotmailer/dotmailer-magento2-extension/issues/502

Related work items: #65828