Merged PR 6368: check and create abandoned carts table
Abandoned cart table is not create in the upgrade script when upgrading from 2.3.8

When upgrading from a bundling version there are no abandoned cart feature and the missing table error for abandoned cart will trigger.

This should fix the error when abandoned cart table was missing inside the recurring script. Recurring file will be executed every time when the setup:upgrade command runs and will check that abandoned cart table should be created on version number 2.3.8 and up.

Assuming that he next versions of the bundle will have the same issue we have to keep on checking everytime the setup upgrade runs.

Related work items: #66277