Merged PR 9252: Support split db for InstallData (and batch email_order population)
This change builds on the uninstall script branch. The PR will be simpler once that branch is merged. The only changes are in InstallData.php.

I've changed it to pull batches out of sales_order into in-memory arrays, then insert back into email_order. This follows the same batching pattern we implemented in M1. batchPopulateTable should be generic enough so that we can use it to populate other tables when we have time.

I'm not sure how you define/import helper 'model' classes like BatchPopulateTableModel? If there's a better way i'm all for it. I briefly looked at annonymous classes in php7 but wasn't convinced it was a nicer solution. Public fields/properties are also a big no no in c# world. Is it ok here? I couldn't see a better way. Please pull this class apart.

Related work items: #72036