Merged PR 10374: Skip keys which will be invalid
Skip keys which will be invalid
Trim first, then run preg_match. If no match, skip.
https://support.dotmailer.com/hc/en-gb/articles/212214538-Using-Insight-data-developers-guide-#restrictkeys

###How to test?

Sync a product with an attribute that's not valid  according to the dotmailer support article. Ensure it's skipped and the product continues to be synced.
Sync a product with valid attributes and ensure its synced in.

Related work items: #77710