Merged PR 11235: Emails could go out with incorrect sender details
In a Magneto instance you can set up multiple stores. Each store can have a distinct email sender. Store X may be set up to send emails from sales@x.com and store Y may be set up to send emaisl from sales@y.com. Both of these emails should exist in dotmailer as custom from addresses. You can also use the dotmailer extension to specify a dotmailer campaign to use for an email template in Magento. Customers reported emails from store Y going out with store X's email sender. It turned out that the emails were in fact going out with a custom from address set in a template that had been synced in from dotmailer.

#### Steps to Reproduce
Set up a dotmailer template mapping from within the backend of magento, let's use new subscriber as an example. Set the CFA for the campaign to be one of your available CFAs, e.g. X.
Set up an email sender in Magento to use one of your other CFAs e.g. Y.
Set up Magento to use dotmailer as an SMTP server.
Send an email that uses the email sender we set up earlier.

#### Actual Result
Email is sent using the CFA in the template (X)

####Expected Result
Email is sent using the one in the email sender (Y)

####Cause
MessagePlugin.php:60
`if ($templateId = $this->isTemplate() && $this->shouldIntercept())` caused `$templateId` to be true which is incorrect.
It should be the template id retrieved from the registry.
Could have bracketed the assignment, but chose to pull out the assignment to separate concerns from the if statement.

Interestingly the unit test could have caught this but the `with(...)` was comparing `true` and `$templateId` which was an integer greater than 0 and deciding the load had been called with the correct values.

Related work items: #80921