Merged PR 11931: Fix limit of 1000 campaigns when selecting 'Campaign to Send' in Abandoned Cart config
Due to the limit of 1000 responses from the `GET /campaigns` API endpoint, customers with > 1000 campaigns cannot select the campaign they wish to apply abandoned cart emails to.

**To test**
Create over 1000 campaigns in EC and check that the most recent is shown in the dropdown (?)
Alternatively, the tests added for this fix can simulate the response.

Related work items: #57867