Merged PR 12189: Find group and bundle product URLs for simple products
This PR attempts to provide reachable product URLs for products that a) are of type Simple, b) are not visible in the catalog, and c) have no configurable parent.

If we cannot find a configurable parent, we look for a group or bundle product ancestry, and if so return the URLs of the group or bundle.

Related work items: #83660