Merged PR 12416: Repair product URLs
WORK IN PROGRESS - CARTINSIGHT STILL TO BE TESTED

Ensure that product URLs are provided correctly for all types of product, in both the catalog sync and during cartInsight data retrieval.

**To test:**

- Adopt a website / store setup similar to

Website 1
 -- store 1
 -- store 2
Website 2 (ad custom domain or subdomain)
 -- store 3

and have some products belong to Website 1, others to Website 2, and others to both websites.

- Run the catalog sync at default level and at store level and verify the checks indicated in the matrix: https://www.dropbox.com/s/qwlbfp301b5s6tq/Product%20URLs.xlsx?dl=0
- All URLs should be correctly rewritten if rewrites are available, and the custom domain configured in website 2 should be evident.