Merged PR 12717: Remove call to getText()
In some rare cases when `$status` was either null or a string (as opposed to an instance of `Magento\Framework\Phrase`), the call to `getText()` was throwing an exception. It's fine to assign the object to `$this->status` because the object's `text` property is used automatically when we `json_encode` the data in preparation for EC import.

Related work items: #81477