Merged PR 13750: Don't fire sync tasks when clicking the label of an item
The 'Reset Sync Options' section under EC > Developer contains a series of buttons which, inconsistently for our Magento config items, have a non-hashed ID set (to match their corresponding `<label>`). As a result of this ID, clicking the label triggers the button action, which could result in some actions being run when the user didn't intend them to be.

This change adds the hashed ID for those 4 buttons, thus rendering a click on the label redundant.

**To test**

- Click the label of the 'Reset Order Data', 'Reset Review Data', 'Reset Wishlist Data' and 'Reset Catalog Data' buttons and confirm the button is not triggered.

Related work items: #63407