Merged PR 13783: Don't mark a Subscriber_Resubscribe job as failed in email_importer if the contact was already subscribed in EC
If a contact is already subscribed in EC, and a Subscriber_Resubscribe job is sent for them, the job is marked as failed with the infamous 'Error Unknown' message.

This change checks the response from EC when sending Subscriber_Resubscribe and, if the contact is already subscribed, the job is marked as complete with the message 'Contact is already subscribed'.

**To test**

- Create new account in Magento, subscribing to newsletter
- Run Subscriber, Contact, Importer syncs
- Suppress in EC
- Run Subscriber sync
- Resubscribe contact in EC
- Run syncs for good measure (no change to contact in Magento - _potential TODO: should this resubscribe in Magneto?_)
- Resubscribe in Magento
- Run Subscriber, Contact, Importer syncs
- Check in `email_importer` table that Subscriber_Resubscribed row is marked as import_status = 2 (complete) and message = 'Contact is already subscribed'

Related work items: #57787