Merged PR 15483: Get template id from first value of $result data
## What's being changed

Our TemplatePlugin script.

## Why it's being changed

In 2.1.x, `dotmailer_current_template_id` is not being saved to the registry, hence in MessagePlugin, the code that obtains the correct EC campaign From name and address is not running.

This is because, in 2.1, $result in TemplatePlugin is a single value at a time, as opposed to a keyed array in 2.2+.

In this PR, we store the first value of $result when TransportBuilder requests the ID of the template for the message at hand.

## How to review / test this change

- Set up 2.1 installation with EC transactional and a handful of email templates configured for different transactional sends.
- Your EC campaign should have a Friendly From name set.
- Run template sync. You should have more than one row in `email_template`.
- Trigger the send.
- The message record in EC transactional outbox should indicate the correct From name from the correct template.

Related work items: #94526