Restrict single brand values
## What's being changed

The method that retrieves the brand value for a given product ID. In the event of multiple values being returned for the brand, we will send them all as comma-separated values, just as we treat multiple categories.

## Why it's being changed

The returned value must be a string. Prior to this change, 'brand' attributes configured as multi-selects were sending arrays to CSV, which caused an error.

## How to review / test this change

- Create a new product attribute for 'Brand'
- Make it a multi-select field
- Apply the attribute to a product and select multiple values
- Set this new brand field to be the 'Brand Attribute' under Configuration > Data fields
- Map FIRST_BRAND_PUR and LAST_BRAND_PUR
- Place an order for the product you added multiple brands to
- Run Contact sync
- You should now see comma-separated brands output in the LAST_BRAND_PUR column

Related work items: #95016