Merged PR 15624: Exit abandoned cart automation if cart is empty
## What's being changed

The flow of the automation sync.

## Why it's being changed

If an abandoned cart automation is triggered, but by the time the automation cron runs, the original cart has been emptied, we do not post a datafield update to EC and instead mark the pending automation as 'Cancelled'.

## How to review / test this change

Part 1
- Configure AC program enrolment
- Map ABANDONED_PROD_NAME
- Drop a cart
- Run the abandoned cart sync to generate a pending automation
- Run the automation cron
- Confirm that the automation is set to 'Active'

Part 2
- Manually reset the 'Active' record back to 'pending'
- Delete all items from your cart
- Run the automation cron
- Confirm that the automation has been set to 'Cancelled' and that data fields have NOT been sent to EC

Bonus
- Check other types of automation e.g. Order

Related work items: #95143