Merged PR 15688: Keep consistent message in syncs
The purpose of this PR is to keep sync messages consistent using this type

----------- Review sync ----------- : 00:00:00, synced = 0

EVEN IF the synced items are 0 .
Previously in most of the syncs we used to Print a message `Done` or even No Message at all if the quantity of synced items were 0.

**How to test**
Simply checkout to this branch and run the syncs and see the messages in both connector.log or on admin panel if you do it via developer buttons

_Note_
Template Sync and Importer sync currently excluded cause there is no implementation for this at all.
(By default the message for them are set to `Done` )

Related work items: #95537