Merged PR 15912: Mark ordered products as modified
https://dotmailer.zendesk.com/agent/tickets/492391

Some clients noticed that they get wrong stock values in catalog sync.

This is happening because after the completion of a product purchase we don't
mark it again as `modified`. this means that every time a user buys a product
then we don't resync the new decreased stock value .

**What changed**
Now during order sync we mark ordered products as modified .
By doing this we re resync these products during catalog sync, so the new stock value will be updated.

**How to test**
Simply complete a purchase of a product that is already imported into EC.
After you run the order sync, then the product should be marked as modified in `email_catalog`

Related work items: #95624