Merged PR 15883: Custom attributes bug fixed
**Repro steps**
https://dev.azure.com/dotdigital/ec/_workitems/edit/96064

**Solution**
So it seems that before the values (in System > Configuration) are going to be saved there is a filter functionality that parsing the system.xml files of all installed modules , then searching the post values that we are trying to save, and if there is a match between parsed values and Post values , only in this case are going to be saved,

So investigated the saving process and i ended up in this Magento's core file
app/code/Magento/Config/Model/Config/Structure.php
In Line 401 there is this piece of code

              if ($element['_elementType'] === 'field' && isset($element['label'])) {
                    $structurePath = (isset($element['path']) ? $element['path'] . '/' : '') . $element['id'];
                    $configPath = isset($element['config_path']) ? $element['config_path'] : $structurePath;

                    if (!isset($result[$configPath])) {
                        $result[$configPath] = [];
                    }

                    $result[$configPath][] = $structurePath;
                }

it seems that isset($element['label']) "rejects" our custom data fields  because we left <label></label> empty.
So i manage to "solve" this case by adding a dot as you ll see in gitdiff.
It seems that is oneway road to add a label.

Related work items: #93322, #96064