Merged PR 16207: Push integration insight data to Engagement Cloud
## What's being changed

We are sending information about Magento integrations to Engagement Cloud when an API is enabled. This includes details of websites connected to a particular API username, and the Magento and connector versions. No personal data is sent.

## Why it's being changed

This will enable advanced functionality in Engagement Cloud to enrich the omni-channel experience for Magento stores and their customers.

## How to review / test this change

1. Once upgraded, run `bin/magento dotdigital:sync IntegrationInsights`
2. Check `var/log/connector.log` and see a reference to integration insight data being sent for each API user enabled in your Magento installation.
3. Check **Insight data** in Engagement Cloud, and confirm that the records listed under the 'Integration' collection are correct.
4. Run an upgrade and confirm that the sync happens
5. Enable an API and confirm that the sync happens

Related work items: #94829