Merged PR 14501: Process product IDs in batches in catalog sync
## What's being changed

The catalog sync is being changed so that it retrieves product IDs to be processed from the `email_catalog` table in advance of syncing by store. We are changing the notion of 'imported' products to 'processed' i.e. on each sync run we **process** 50 (if that is the batch size) products, of which we may only **sync** 5. By marking all 50 as processed, however, we ensure that we don't check them again on the next sync run.

## Why it's being changed

Some clients only sync a subset of products (e.g. configurable, visible products), and in these cases - especially with small sync batch sizes - it is possible for the sync to run out of qualifying products.

## How to review / test this change

- First, upgrade

*Catalog sync*
- Configure the catalog sync to only sync a sub-set of products e.g. configurable, visible in catalog and search
- Reset the catalog
- Run the sync
- Test at store level and default level
- Test a multi-store setup, e.g. with products included in the sync that belong to one or other store, or both.
- Set the batch size to 500, and let the cron run until all products are processed.
- Test modifying an already-processed product and ensure that still works.

*Observe*
- Check the log messages in connector.log
- Check the `email_catalog` table
- Check the `email_importer` table
- Check Engagement Cloud insight data

*Additional tests*
- Test single product update
- Test bulk product update
- Test reset catalog
- Check the grid at /admin/dotdigitalgroup_email/catalog/index/

*Schema modifications*
- Test upgrade changes `email_catalog` imported column to 'processed'
- Test install sets `email_catalog` 'processed' column correctly
- Check 'modified' column is removed
- Check 'last_imported_at' column is created
- Check index is set on processed column and named correctly

Related work items: #91371