Merged PR 16382: Hotfix: confirm catalog sync is active at some level before processing
Now that we gather products to process before passing them to either the default level catalog syncer or the store level catalog syncer, we need to avoid processing unnecessarily if sync is disabled.

This PR ensures that, if catalog sync is disabled everywhere, no products will be processed, and no logs output.

It is still possible that processing happens unintentionally in a misconfigured environment. Consider the following configuration (assume that sync level is set to Default):

| Scope | Catalog sync enabled |
| :-------------: |:-------------:|
| Default level | No |
| Website 1 | No |
| Website 2 | Yes |

Now, since catalog sync is enabled for website 2, products will be processed, but because sync level is set to Default, the script is routed to the default level catalog syncer, and therefore continues to process products without syncing any of them.