Merged PR 16771: prevent XSS in Marketing exclusion rules
https://jira.corp.magento.com/secure/RapidBoard.jspa?rapidView=1564&projectKey=BUNDLE&view=detail&selectedIssue=BUNDLE-2159&quickFilter=6486

**Steps To Reproduce:**
1. Click on `Marketing->Exclusion Rules` and then click on "Add New Rule"
2. Enter arbitrary non-ambiguous rulename, and select arbitrary website
3. Click on the Conditions tab, and click on "Add New Condition"
4. Select Attribute to "Email", Condition to "is" (arbitrary) and Value to "" (leave empty, arbitrary).
5. Now you have to edit the Form which is sent to the server:
For each of the three `<input>`/`<select>` fields (Attribute, Condition and Value)
Rightclick -> inspect element, and modify the "name" field of the HTML element
from something like `condition[_1562168869914_914][attribute]` to `condition['><script>alert(1)</script><i a='][attribute]`. (leave the second array key of condition as it is!)
6. Now click on save, you should retrieve a "successfully saved" message
7. As soon as someone clicks on "Edit" next to the newly created url, the XSS is triggered.

_Impact_
A lower privileged backend user can use this XSS to perform actions in the name of a higher privileged admin user.

**What changed**
A new `preventXSS` function added to prevent attacks in admin area

Related work items: #98941