Merged PR 17436: Catch exceptions when requesting non-existent parent products
## What's being changed

The URL Finder utility and associated unit tests.

## Why it's being changed

It is possible on some installations for parent IDs to be related to a child, for those parent IDs not to correspond with an actual product. Hence we are try/catching the `load()` call from the product repository, to enable the sync to proceed even if an exception is thrown.

## How to review / test this change

- In UrlFinder L:245, pass in a broken product reference

To break the product either:
- replace getFirstParentId() with a fake id
- delete a parent Product via db (not via magento admin panel)

Related work items: #101301