Merged PR 17634: Determine price for order line items
## What's being changed

Price logic for order line items in order sync.

## Why it's being changed

For configurable products, we now use the configurable line item price as the price for the child - because if either a) custom options are set which have a cost attached or b) the simple child or children have a different price from their parent, the configurable line item price reflects the actual price paid.

For all other types of product, we use the line item price.

## How to review / test this change

- Add a collection of configurable and bundle products to the cart
- Ideally you want an order with simple line items that show 0 for `price` in the table, AND line items where the price varies between sibling children of a common parent.
- Ensure price is never 0 in the synced data.

Related work items: #101830