Merged PR 17689: Remove retrieval of store id from newsletter_subscriber table
## What's being changed

We're changing the way we source a store id for a given set of contacts in the SubscriberExporter class.

## Why it's being changed

When exporting subscribers for all stores, we have been retrieving a store id from each contact's record in the `newsletter_subscriber` table - and if we don't find a match, defaulting to store id 1. This actually pre-dates a change we made in 3.2.2 to batch subscribers by store instead of by website.

Either way this approach was flawed because:
1. The store id in newsletter_subscriber might not be valid, resulting in a `Requested store is not found` exception.
2. Merchants may not have a store with id 1 (again leading to an exception).
3. We were making unnecessary db queries.

This PR changes the logic so that the store id is taken from the store for which we are currently exporting subscribers. This removes the queries and the extra `storeManager` load.

## How to review / test this change

- Subscribe a handful of contacts from different stores
- Run contact sync
- Check the csv data is correct

Related work items: #101852