Merged PR 17825: Adjust DOB for timezones between -1 and -11
## What's being changed

How we return the `dob` value from the customer entity.

## Why it's being changed

Engagement Cloud adjusts incoming dates according to the timezone of the account. For customers being imported from Magento into an EC account in a timezone between -1 and -11, therefore, birthdays of e.g. 1/1/1980 were being stored as 31/12/1979.

This PR adjusts the date (according to the timezone set in Magento) by adding the same amount of time that is due to be subtracted in EC - resulting in the correct birthday. For example 1/1/1980 (for an LA store at GMT -08:00] is sent as 1980-01-01T08:00:00+00:00.

This does rely on the Magento store locale matching the EC timezone.

## How to review / test this change

- Set the timezone of your EC account (Account > Account settings > My timezone) to a GMT minus timezone.
- Set your Magento locale to the same timezone (General > Locale Options).
- Log in as a customer and set your date of birth.
- Run Contact sync (you'll need datafields to be mapped already)
- Run the importer
- Check your contact record in EC and confirm that your birthday is correct.

Related work items: #101786