Merged PR 18571: Remove escapeUrl in trial and embed blocks
## What's being changed

Usages of `escapeUrl()` in trial and embed JS blocks are being removed.

## Why it's being changed

We previously used `escapeUrl` to escape some data in the trial and embed JS blocks. However this changes `&` to `&amp;` and thus breaks the iframe.

`escapeJs()` does work as a replacement, and indeed `encodeUrlParam()` could be used for the constructed URL params, but we're removing it completely for now lest this have some other unexpected impact.

## How to review / test this change

- Clear EC creds
- Start a trial for either EC or Chat
- Check the trial completes as expected
- Confirm the EC embed works once setup is complete

Related work items: #105175