Merged PR 18933: Remove PHP unserialize() from upgrade script
## What's being changed

The upgrade script for merchants on < 2.2.1 of our connector.

## Why it's being changed

This PR removes the last remaining call to PHP's native unserialize() method, which is discouraged by Magento. It was still in place to convert M1-style serialized strings stored in several of our tables, into JSON - but this was an upgrade written for the 2.2.1 release in April 2017.

Merchants who are still on older versions of our connector than 2.2.1 are advised to upgrade, then run our migration script to refresh the data in our `email_` prefixed tables.

```
bin/magento dotdigital:migrate
```

## How to review / test this change

- Set your connector version to < 2.2.1
- Run setup:upgrade
- Check that the remaining parts of the upgrade script for this version are run without error.

Related work items: #101528