Merged PR 19075: Allow individual coupons to have an expiry date
## What's being changed

This PR adds an optional expiry date to an individual coupon, which will be validated in addition to the sales rule it belongs to at checkout.

## Why it's being changed

This enables connector users to continue to set an expiry date for coupons generated via EDC, following Magento's decision to deprecate this functionality from core.

## How to review / test this change

- Create a Sales Rule. This rule MUST:
-- Be for a specific coupon
-- Use auto-generation
-- Have no 'To' date set (if this is set, the 'Expires After' field in the coupon builder is greyed out)

- Use the dotdigital coupon builder to generate an EDC URL containing an expiry date
- Generate a code
- Confirm that the correct `expires_at` datetime is shown in the `email_coupon_attribute` table (x days specified from the current day, in UTC)
- Confirm that the date appears in the 'Manage Coupon Codes' grid. The original Expiration Date column has been removed (it appears we were adding that!)
- Apply the coupon to a quote at checkout
- Confirm the coupon is applied
- Cancel the coupon
- Edit the expiry date in `email_coupon_attribute` to be in the past
- Reapply the coupon and confirm it is rejected

## Notes

Coupon expiry date is stored as UTC, but will be converted to the active locale of the website it is being redeemed in when being applied to a quote. Some additional testing should be carried out when resetting the expiry and checking whether the coupon is valid.

Related work items: #106065