Merged PR 19545: Check that DDG extension attributes exist before checking for expiry date
## What's being changed

We are adding an additional check for extension attributes before the expiry date check.

## Why it's being changed

Redeeming coupons generated via the connector in versions prior to 4.3.0 were causing a fatal error in version 4.3.4.

## How to review / test this change

- Create a coupon code via EDC/coupon builder
- Remove it's corresponding row from the email_coupon_attribute table (this simulates an EDC coupon generated prior to
4.3.0)
- Attempt to redeem the coupon at checkout
- Confirm no fatal error occurs

## Notes

The Coupon Codes URL field has been changed to a textarea, as this means you can see the changes to the EDC url as they are made.

Related work items: #109025