Merged PR 19579: Conditional add b2b reset quote css selector
Currently in daterange js we include css selectors of (some) of the run developers button to be mapped in "from" and "to" filter.
In new b2b module we added a new button that is not mapped.
So now we add a conditional (if b2b module is enabled or no) inclusion of this selector.

**How to test**
- Checkout to b2b's module quote develop branch
- Go to EC > Configuration > Developer
- Select From and To dates and run B2b reset quotes button
- Open network tab and confirm that from and to params included in post action
- Disable b2b module choose again from and to dates and run the order reset button. Make sure it works as expected without js errors in console.

Related work items: #109148