Merged PR 19426: Enable processing of attribute_set_id conditions in exclusion rules
Enable processing of attribute_set_id conditions

## What's being changed

Exclusion rule logic.

## Why it's being changed

Conditions set for 'Attribute set' were not functioning. Attribute set is not a product attribute per se, nor is it one of our quote- or order-specific attributes. Hence it wasn't being processed when the form was submitted. This PR fixes that.

## How to review / test this change

- Drop a cart containing e.g. a Bag and a Top
- Set up an Abandoned Cart exclusion rule with a condition to exclude products with an Attribute Set equal to 'Gear'
- Run the AC sync
- You should see a row in email_abandoned_cart
- Now remove this row
- Edit your rule and set the value to 'Bag'
- Re-run the AC sync
- No row should be created in the table
- Experiment further with different conditions and combinations

## Notes

The attribute set list may contain duplicates. This reflects the `eav_attribute_set` table, which lists attribute sets with the same name but different `entity_type_id` values. I don't altogether understand this. I _could_ exclude the duplicates but then what if I exclude an `attribute_set_id` that a merchant really wants to filter on? I expect that merchants will tend to want to use this condition to filter out orders (or quotes) with products belonging to custom attribute sets, which I predict will not be duplicated - so I've left it as is for now.

Related work items: #71107