Merged PR 19802: Check that bodyPart is of matching type before calling setEncoding
## What's being changed

We're adding an extra condition to the beforeSetBody method of our MessagePlugin.

## Why it's being changed

The code to override the encoding on pre-assembled mime messages was [introduced in 4.3.0](https://github.com/dotmailer/dotmailer-magento2-extension/commit/ed9d1a93f7d3437f351ceef117a9e3a03ed29ad5), specifically to fix a problem with a third-party module, in Magento 2.2.

In other, more up-to-date instances, it may cause an error because it is possible for a `Zend\Mime\Message` to be composed of parts of type `Magento\Framework\Mail\MimePart` - and this class has no `setEncoding()` method.

## How to review / test this change

The only way to properly test this code would be to install (or create) a third-party module which calls `Magento\Framework\Mail\MessageInterface::setBody()`. Without this, simply validating the code on sight should be enough.

Related work items: #110058