Merged PR 19771: Fix MFTF tests for coupon expiry and address book mapping
## What's being changed

This fixes two of our tests.

## Why it's being changed

For the coupon test, we're testing interactions with the revised 'Expires after' field in the coupon URL builder tool. For address book mapping, there's a small test to add a random suffix to the created address book names, which ensures the suite can be re-run and not fail because of creating duplicate address books.

## How to review / test this change

- Run the affected tests. Add a custom group e.g. 'fix' to the modified tests and run `vendor/bin/mftf run:group fix -r`

Related work items: #108924