Merged PR 20098: Reset account data after coupon MFTF test
## What's being changed

The coupon URL builder MFTF test.

## Why it's being changed

We must clear EC account credentials after each test that sets them in the before step.

If this isn't done, depending on the order the test suite is run in, we leave the account active in advance of another test that may only succeed if the account fields are in a default (blank) state.

## How to review / test this change

- Run the MFTF test

Related work items: #110898