Merged PR 20206: Remove check for items_qty in automation enrolment
## What's being changed

For abandoned cart program enrolments,

a) we now send cartInsight data if a quote is active but has no items
b) we now use the same Abandoned Cart Limit value as the AC series, to block re-enrolment inside that time limit.

## Why it's being changed

Change a) means that once a customer drops a cart and is enrolled in a program, if they then clear their cart, we will send cartInsight data with 0 items. This in turn can be used as a logic gate in a program to ensure a customer is not sent further email if their cart is empty.

Change b) means that if you set Abandoned Cart Limit to 2 hours, a customer won't get re-enrolled to the same program if the sync runs and finds an updated cart in the last 2 hours. Note that cartInsight will always be sent.

## How to review / test this change
- Drop a cart in the basket
- wait till you are enrolled to automation and let the automation begin
- once the automation is active log in again and empty your cart
- wait for the next automation cron to run see the updated cartInsight data and make sure that we do not have any line items

Related work items: #110814