Merged PR 20478: Enable sync of store group name
## What's being changed

Added a new data mapping field and a new method to grab the REAL store name aka the 'Store Group Name'.

## Why it's being changed

Currently in STORE_NAME data field we're mapping the store VIEW name. so we need to find a way to map the REAL store name as well

## How to review / test this change

- Go to Data mapping section and make sure that "Store Name" label changed into "Store View Name"
 also make sure that a new Data Field (unmapped) added with label "Store Name"

- Create a new Data Field and assign in to new "Store Name" field that we recently added

- Run Contact sync and confirm that we sending the REAL store name

Related work items: #111937