Merged PR 20525: Improving consistency of our logs
## What's being changed

We're improving the quality and consistency of log lines output by the `Apiconnector\Client` class.

## Why it's being changed

We had started this work before, but were asked by Magento architects to wait until a minor Magento release before changing the log format. With 2.4 release on the horizon, it's time to release these improvements.

## How to review / test this change

- Pls see test steps / notes on this original PR https://dev.azure.com/dotdigital/ec/ec%20Team/_git/ec-core-magento2-extension/pullrequest/16009

Related work items: #112056