Merged PR 20421: Fix wishlist sync logs and admin message
## What's being changed

This is a small refactor of the wishlist sync.

## Why it's being changed

This fixes the inaccurate log lines output by this sync (both to log file and to screen).

## How to review / test this change

- Create wishlists for a customer in website 1 and 2
- Run the sync from the CLI, check the logs
- Mark one of the wishlists as modified, re-run the sync
- Mark one of the wishlists as not imported, re-run the sync
- Reset wishlists
- Run the sync from the admin, observe the shiny new success message

## Notes

We should be going further here, and removing the 'single' sync stage by pushing all unprocessed or modified wishlists into the importer for bulk upsert to EC.

Related work items: #108801