Merged PR 20687: Improve check for missing status in resubscribe response
## What's being changed

Our handling of the response we receive back from a resubscribe API call.

## Why it's being changed

`if (!$response->status) {}` throws an exception if `$response` is an object containing just a `message`. It's possible we get this back from either resubscribe call (either to an address book or just to the account in general).

This was not a regression from the recent importer refactor in 4.3.3, it seems we always did a false-y check on `$response->status`. Getting a message back from the resubscribe API calls must be rare.

## How to review / test this change

- In order to simulate a response without a status, we have to add lines above line 234 in /Model/Sync/Importer/Type/Contact/Update.php:
```
$response = new \StdClass;
$response->foo = "bar";
```
- Now generate a`Subscriber_Resubscribed` importer row
- Run the importer and see we don't get an exception

Related work items: #113040