Merged PR 21759: Return false if last_subscribed_at is null
## What's being changed

We're adding an additional check on the `last_subscribed_at` column before deciding to resubscribe contacts.

## Why it's being changed

If there's a suppression for an email address on EC, and that email address is then imported again in another context (an order for example), EC will add the address to the import report fault. We digest the report after every import, and check if any of the listed contacts were resubscribed since the last cron ran. If we find any, we resubscribe them.

```
new \DateTime($contact['last_subscribed_at'], new \DateTimeZone('UTC'))
```

The problem is if we pass `null` as the date value, this defaults to now, which is always greater than the last sync time. The result in some cases is that suppressed contacts are continually resubscribed.

## How to review / test this change

- Suppress a contact in EC
- Resubscribe that contact in Magento
- Clear the `last_subscribed_at` value from the row in `email_contact`
- Run subscriber sync
- Run importer sync
- You should see the contact in the contact import report faults csv (if not, add dummy csv to `_processContactImportReportFaults()` like:

```
$report = 'Reason,email,firstname,fullname,gender,lastname,postcode
Suppressed by you,test-host@emailsim.io,Chaz,Kangeroo,Male,Kangeroo,123456';
```
- Check that no Subscriber_Resubscribe is created for test-host@emailsim.io.

Related work items: #116899